backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.123.10",
    "openai-agents>=0.6.2",
    "openai-chatkit>=1.4.0",
    "python-dotenv>=1.2.1",
]
